Faster PNG striping
I found that when the image is really big the ChunkyPNG gem is slow to
encode/decode the PNG pixels. As we only want to strip the metadata we
can work in the Datastream that doesn't transform the pixels.

With my test image this method went from 10s to 0.01s.